+++ src/assetbundles/transcoder/src/js/Transcoder.js
@@ -1,6 +1,6 @@
 // Import our CSS
-import styles from '../css/app.pcss';
-import icon from '../img/Transcoder-icon.svg';
+import '../css/app.pcss';
+import '../img/Transcoder-icon.svg';
 
 const main = async () => {
 };